Add files via upload
This Python script solves the 'Hanging Cable Problem', allegedly featured as an interview question at Amazon. Given a cable of 80m in length, suspended between two towers of 50m in height, hanging 20m above the ground at its lowest point, the candidate is asked for the distance between the two poles. The question is repeated with the cable hanging only 10m above the ground. (Hint: at 10m, the cable is neatly folded in two as the towers are at the same point.) 

This script prompts the user to enter the following data:
1. The length of the cable (in meters),
2. The height of the first tower (in meters),
3. The height of the second tower (in meters), and
4. The height of the cable at its lowest point (in meters). If the towers are of unequal height, then enter 0.

If the towers are of equal height, then we use the Hyperbolic Identity (cosh**2(t) - sinh**2(t) = 1) to determine the distance between the towers. If the towers are of unequal height, then the math gets a lot uglier, eventually using the Quadratic Formula to determine the distance between the towers. All results are rounded to 3 decimal places.